Rename pal_stat.h/cpp to pal_io.h/cpp
Rationale: grouping related functions that need to share constants
